+++ rusoto/services/guardduty/src/lib.rs
@@ -1,0 +1,33 @@
+
+// =================================================================
+//
+//                           * WARNING *
+//
+//                    This file is generated!
+//
+//  Changes made to this file will be overwritten. If changes are
+//  required to the generated code, the service_crategen project
+//  must be updated to generate the changes.
+//
+// =================================================================
+
+#![doc(html_logo_url = "https://raw.githubusercontent.com/rusoto/rusoto/master/assets/logo-square.png")]
+//! <p>Assess, monitor, manage, and remediate security issues across your AWS infrastructure, applications, and data.</p>
+//!
+//! If you're using the service, you're probably looking for [GuardDutyClient](struct.GuardDutyClient.html) and [GuardDuty](trait.GuardDuty.html).
+
+extern crate futures;
+#[macro_use]
+extern crate log;
+extern crate rusoto_core;
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate serde_json;
+
+mod generated;
+mod custom;
+
+pub use generated::*;
+pub use custom::*;
+